requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
openpyxl==3.1.2
